Add support for Yan
Yan is disabled by default. It can be enabled with an option
in the settings.

This is quite experimental, but it should not bother current users
since nothing change until the package settings are modified.

I have removed the node_modules/ directory. It must not be
committed with Git. Use 'npm install' to resolve the dependencies
locally.